refactor: refactor queue to mirror changes to counter implementation
