+++ src/usecases/gas_killer/ingress.rs
@@ -1,19 +1,55 @@
 #![allow(dead_code)]
+use crate::usecases::gas_killer::creator::{SimpleTaskQueue, TaskQueue};
+use alloy_primitives::{Address, U256};
 use axum::{Json, Router, extract::State, http::StatusCode, routing::post};
-use std::sync::{Arc, Mutex};
+use serde::{Deserialize, Serialize};
+use std::sync::Arc;
 use tracing::info;
 
-use crate::usecases::gas_killer::structs::{GasKillerTaskRequest, GasKillerTaskResponse};
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct GasKillerTaskRequestBody {
+    pub target_address: Address,
+    pub call_data: Vec<u8>,
+    pub storage_updates: Vec<u8>,
+    pub transition_index: u64,
+    pub from_address: Address,
+    pub value: U256,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct GasKillerTaskRequest {
+    pub body: GasKillerTaskRequestBody,
+}
+
+#[allow(dead_code)]
+impl GasKillerTaskRequest {
+    pub fn is_valid(&self) -> bool {
+        let body = &self.body;
+        if body.target_address.is_zero()
+            || body.call_data.is_empty()
+            || body.storage_updates.is_empty()
+            || body.transition_index == 0
+        {
+            // TODO: add additional checks
+            return false;
+        }
+        true
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct GasKillerTaskResponse {
+    pub success: bool,
+    pub message: String,
+}
 
 // Handler for POST /trigger
 pub async fn trigger_task_handler(
-    State(state): State<Arc<Mutex<Vec<GasKillerTaskRequest>>>>,
+    State(queue): State<Arc<SimpleTaskQueue>>,
     Json(request): Json<GasKillerTaskRequest>,
 ) -> (StatusCode, Json<GasKillerTaskResponse>) {
-    if let Ok(mut queue) = state.lock()
-        && request.is_valid()
-    {
-        queue.push(request.clone());
+    if request.is_valid() {
+        queue.push(request);
         return (
             StatusCode::OK,
             Json(GasKillerTaskResponse {
@@ -32,10 +68,7 @@ pub async fn trigger_task_handler(
 }
 
 // Start the HTTP server in a background task
-pub async fn start_gas_killer_http_server(
-    queue: Arc<Mutex<Vec<GasKillerTaskRequest>>>,
-    addr: &str,
-) {
+pub async fn start_gas_killer_http_server(queue: Arc<SimpleTaskQueue>, addr: &str) {
     let app = Router::new()
         .route("/trigger", post(trigger_task_handler))
         .with_state(queue);
